feat: DynamoDB database migration/enrichment
To allow performant listing of submissions, the dynamodb table
needs to be enriched with more information. This PR (in progress)
handles that migration.

We add a new Migration class, which performs a table scan, then grabs
info from S3, merges this data and updates the table items. The
included test uses a local dynamodb table to perform this update with
a fictional dataset. To run this, it assumes:
- you have docker running
- There's a .env in the root of the project with at least the contents:
`JEST_RUN_INTEGRATION_TESTS = TRUE`

You probably need to run `npx projen` since this adds some (dev)
dependencies.