feat: Feature flag authorizer
Puts the authorizer functionality, which allows
authorization using (JWT) tokens behind a feature
flag. If the flag is off, the authorizer will not
be added to the api gateway routes, and the lambda
will request identifers from the query params
instead of receiving them from the request context
via the authorizer.