feat: Update listing and getting submissions by pk
The recent migration fixing organisation submissions (by kvk) changes
the PK format (to ${PERSON|USER|ANONYMOUS}#hash). This commit propagates
using that format to the list and get endpoints for the http-API.

The API already requires sending a userType in addition to a userId, so
this change should be backwards compatible for all consumers of those
endpoints. Since we currently have doubled records in the DB table, end-
points requesting records without using the PK may not works as expected